Remove CFLAGS from tests/Makefile.am
Test are intended to be run from the root makefile, which delegates to
tests subfolder. If tests define their own CFLAGS, any configuration
from the root is overridden. This may result in a situation where the
tests fail to execute, because the library was compiled with
incompatible flags.

In particular, the starard LDFLAGS used by Fedora are incompatible with
the tests subfolder CFLAGS. Since there does not appear to be any
particular reason for forcing the tests subfolder CFLAGS, they are
removed here.